单品码绑定属性接口新增，产品码属性新增返回最新id修改

+++ acc-module/src/main/java/com/ztl/gym/code/service/ICodeAttrService.java
@@ -41,7 +41,7 @@ public interface ICodeAttrService {
      * @param codeAttr 码属性
      * @return 结果
      */
-    public int insertCodeAttr(CodeAttr codeAttr);
+    public Long insertCodeAttr(CodeAttr codeAttr);
 
     /**
      * 修改码属性
